Rewrite to use custom Graph object
We switch to a custom Graph object which attempts to take care of all
the re-writing of the commit graph after any changes have taken place.

It is an early version and pretty rough around the edges but it wraps
the commit graph walking and return wrapped commits which can have their
messages edited and the changes will be tracked and written into to the
repository on request.

Lots of ugly little bits. We always write new commits even when nothing
has changed. We don't have a clean way to return the new head of the
branch. Other little things like that.

Makes for a nice API as far as being able to change a commit message and
write it all to disk.