Add a recursive 'changed' check
We don't want to write out new commits for commits which haven't
changed. We implement a check to see if a node has any overrides or if
its parents have changed at all. If a parent has changed then the node
has changed by default.

It is expensive to keep tracking back over the network for every single
node. We should cache it somewhere but not on the node itself for the
moment. Probably in the write method of the graph somehow.