Switch to a single parents array
Instead of having touched and untouched parents we combine them into a
single or ordered dictionary and start off the parents as ReferenceNodes
that don't need to be written and replace any of them that we hit in the
graph walking with actual Node instances.

We use an OrderedDict so that the parents order is maintained (which was
a problem with separate arrays) and we use a dict-type instead of a
list-type so that it is easy to replace the ReferenceNodes with actual
Nodes when needed.